requirements.txt
langchain
langchain-openai
langchain-community
requests
python-dotenv
ipykernel
google-search-results
langgraph